Extracted fromPostgresVerbose and fromPostgres into own file.
